% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPreParas.multi.R
\name{estPreParas.multi}
\alias{estPreParas.multi}
\title{Estimate Required Parameters from a Pilot Data (Multiple Cell Types)}
\usage{
estPreParas.multi(
  counts,
  cell.info,
  id = "SampleId",
  x1 = "condition",
  cellcluster = "cellcluster",
  cells.interesting = NULL
)
}
\arguments{
\item{counts}{Normalized count matrix.}

\item{cell.info}{Cell information.}

\item{id}{Sample/subject ID.}

\item{x1}{Groups/conditions/stages.}

\item{cellcluster}{cell types.}

\item{cells.interesting}{Cell types of interest.}
}
\description{
A function to estimate required parameters from a pilot data for scPS.
}
\examples{
# https://github.com/cyhsuTN/scPS
# load(file = "DataForDemo/COVID19n.rda")
# counts <- COVID19n$counts
# cell.info <- COVID19n$cell.info
# geneObject <- estPreParas.multi(counts, cell.info,
#                     id="SampleId", x1="condition",
#                     cells.interesting=c("T cells", "DC", "Prolif.T")[c(2,3)])
# https://github.com/cyhsuTN/scPS
# load(file = "DataForDemo/GSE120575n.rda")
# counts <- GSE120575n$counts
# cell.info <- GSE120575n$cell.info
# cell.info$TX <- factor(cell.info$TX, levels = c("Pre", "Post"))
# geneObject <- estPreParas.multi(counts, cell.info, id="ptID", x1="TX",
#                  cells.interesting=c("CD8", "CD4", "Macrophage", "NK", "B")[4:5])

}
