% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPreParas.R
\name{estPreParas}
\alias{estPreParas}
\title{Estimate Required Parameters from a Pilot Data (a Single Cell Type)}
\usage{
estPreParas(counts, cell.info)
}
\arguments{
\item{counts}{Normalized count matrix.}

\item{cell.info}{Cell information.}
}
\description{
A function to estimate required parameters from a pilot data for scPS.
}
\examples{
# https://github.com/cyhsuTN/scPS
# load(file = "DataForDemo/COVID19n.rda")
# counts <- COVID19n$counts
# cell.info <- COVID19n$cell.info
# cells.interesting <- which(cell.info$cellcluster=="Prolif.T")
# counts.small <- counts[,cells.interesting]
# cell.info.small <- data.frame(id=cell.info[cells.interesting, "SampleId"],
#        x1=ifelse(cell.info[cells.interesting, "condition"]=="Neg", 0, 1) )
# aa <- estPreParas(counts=counts.small, cell.info=cell.info.small)

}
