% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSdmean.R
\name{fitSdmean}
\alias{fitSdmean}
\title{Fit a Curve between Gene Means and Gene Standard Deviations}
\usage{
fitSdmean(
  preParas,
  sp.method = "ns",
  probs = c(0.33, 0.67),
  span = 0.75,
  cell.name = NULL
)
}
\arguments{
\item{preParas}{estPreParas object.}

\item{sp.method}{Spline methods: "bs" (cubic B-spline), "ns" (natural cubic spline),
and "loess" (LOESS smoothing). Default = "ns".}

\item{probs}{Inter quantile that will be used as knots in "bs" and "ns" methods.}

\item{span}{A parameter which controls the degree of smoothing if sp.method = "loess".}

\item{cell.name}{cell name.}
}
\description{
A function to fit a curve between gene means and gene standard deviations.
}
\examples{
# https://github.com/cyhsuTN/scPS
# load(file = "DataForDemo/COVID19n.rda")
# counts <- COVID19n$counts
# cell.info <- COVID19n$cell.info
# cells.interesting <- which(cell.info$cellcluster=="Prolif.T")
# counts.small <- counts[,cells.interesting]
# cell.info.small <- data.frame(id=cell.info[cells.interesting, "SampleId"],
#        x1=ifelse(cell.info[cells.interesting, "condition"]=="Neg", 0, 1) )
# aa <- estPreParas(counts=counts.small, cell.info=cell.info.small)

# ## Select 2000 candidate genes
# idx.selected <- which(aa$nonZeroPs1>0.1 & aa$nonZeroPs2>0.1)
# logFC <- log(aa$means2[idx.selected]/aa$means1[idx.selected])
# idx.cg <- head(idx.selected[order(abs(logFC), decreasing = T)],2000)
# aaP <- aa[idx.cg,]
# curve.view <- fitSdmean(preParas=aaP, sp.method=c("bs","ns","loess")[2])
# hf <- curve.view$hf.sigma
}
