% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geeDEA.R
\name{geeDEA}
\alias{geeDEA}
\title{DEA Using GEE for Independent Two-Group Comparison}
\usage{
geeDEA(counts.test, cell.info, hf = NULL, BCC = 1)
}
\arguments{
\item{counts.test}{Count matrix of tested genes.}

\item{cell.info}{Cell information (patients' id and x1 are required).}

\item{hf}{A curve showing the relationship between gene means and gene deviations.
The default = NULL, assuming there is no relationship between
gene means and gene deviations, and gene variances of group 1 are the same as
those of group 2.}

\item{BCC}{Bias covariance correction at small sample sizes. = 0, no correction;
= 1 (Default), t distribution with a degree of freedom = (total size) - 2; = 2,
t distribution and MD's bias covariance correction.}
}
\description{
A function to perform differential expression analysis using GEE
for independent two-group comparison.
}
\examples{
# https://github.com/cyhsuTN/scPS
# load(file = "DataForDemo/COVID19n.rda")
# counts <- COVID19n$counts
# cell.info <- COVID19n$cell.info
# cells.interesting <- which(cell.info$cellcluster=="Prolif.T")
# counts.small <- counts[,cells.interesting]
# cell.info.small <- data.frame(id=cell.info[cells.interesting, "SampleId"],
#        x1=ifelse(cell.info[cells.interesting, "condition"]=="Neg", 0, 1) )
# aa <- estPreParas(counts=counts.small, cell.info=cell.info.small)

# ## Select 2000 genes that will be tested
# idx.selected <- which(aa$nonZeroPs1>0.1 & aa$nonZeroPs2>0.1)
# logFC <- log(aa$means2[idx.selected]/aa$means1[idx.selected])
# idx.cg <- head(idx.selected[order(abs(logFC), decreasing = T)],2000)
# aaP <- aa[idx.cg,]
# curve.view <- fitSdmean(preParas=aaP, sp.method = "ns")
# hf <- curve.view$hf.sigma

# ## Test the selected genes
# counts.test <- counts.small[idx.cg,]
# geeout <- geeDEA(counts.test, cell.info=cell.info.small, hf=hf, BCC=1)
# head(geeout)
}
