% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCandidate.R
\name{geneCandidate}
\alias{geneCandidate}
\title{Select Candidate Genes and DEGs}
\usage{
geneCandidate(
  geneObject,
  cells.interesting = NULL,
  thres.nonZeroPs = 0.1,
  nGenesCandidate = 2000,
  propDEGs = 0.01,
  FC.way = c("median", "individual")[1],
  cg.name = NULL,
  ...
)
}
\arguments{
\item{geneObject}{Object of estPreParas.multi.}

\item{cells.interesting}{Cell types of interest.}

\item{thres.nonZeroPs}{A threshold of non-zero proportion to exclude non-expressed genes.}

\item{nGenesCandidate}{Number of candidate genes (pre-specified). Genes with large observed
fold-changes are selected as candidate genes.}

\item{propDEGs}{A proportion of DEGs among candidate genes (pre-specified).}

\item{FC.way}{FC.way = "median" (default) assumes same logFC = \emph{d}*sign(logFC)
for each DEG, where \emph{d} is the median of |logFC|'s among DEGs;
FC.way = "individual" assumes logFC of each DEG = its observed log ratio of means between groups.}

\item{cg.name}{Names of candidate genes if users want to select candidate genes by themselves.
The default = NULL, meaning genes with large observed fold-changes are selected as candidate genes.}

\item{...}{Arguments in fitSdmean: \emph{sp.method}, \emph{probs}, and \emph{span}.}
}
\value{
\item{logFC.cg}{Observed logFCs for candidate genes}

\item{alogFC.deg}{Median of |logFC|'s among DEGs}

\item{FC}{Assumed FCs for power and sample size calculation}

\item{icc}{ICCs for candidate genes}

\item{hf}{SD-Mean relationship function}

\item{idx.cg}{Indices of candidate genes}

\item{idx.deg}{Indices of DEGs}
}
\description{
A function to select candidate genes and DEGs for power and sample size calculation.
}
\examples{
# https://github.com/cyhsuTN/scPS
# load(file = "DataForDemo/COVID19n.rda")
# counts <- COVID19n$counts
# cell.info <- COVID19n$cell.info
# geneObject <- estPreParas.multi(counts, cell.info,
#                     id="SampleId", x1="condition",
#                     cells.interesting=c("T cells", "DC", "Prolif.T")[c(2,3)])
# Genes.tested <- geneCandidate(geneObject, FC.way="median")
# https://github.com/cyhsuTN/scPS
# load(file = "DataForDemo/GSE120575n.rda")
# counts <- GSE120575n$counts
# cell.info <- GSE120575n$cell.info
# cell.info$TX <- factor(cell.info$TX, levels = c("Pre", "Post"))
# geneObject <- estPreParas.multi(counts, cell.info, id="ptID", x1="TX",
#                  cells.interesting=c("CD8", "CD4", "Macrophage", "NK", "B")[4:5])
# Genes.tested <- geneCandidate(geneObject, FC.way="median")
}
