% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalCost.R
\name{optimalCost}
\alias{optimalCost}
\title{Optimal Combination of m and n}
\usage{
optimalCost(
  m.n.power,
  costfun = function(m, n) 0.24 * m * n,
  ePower = 0.8,
  budget = NULL,
  prop.this.cell = NULL
)
}
\arguments{
\item{m.n.power}{Object of sizeCal/sizeCal.multi or sizeCal.BA/sizeCal.multi.BA.}

\item{costfun}{Cost function. For example, costfun = function(m, n) m*n.}

\item{ePower}{Expected overall power.}

\item{budget}{Budget. The default = NULL.}

\item{prop.this.cell}{Proportion of this cell type of interest. The default = NULL.}
}
\description{
A function to show optimal m and n combination given a cost function.
}
\examples{
# https://github.com/cyhsuTN/scPS
# load(file = "DataForDemo/COVID19n.rda")
# counts <- COVID19n$counts
# cell.info <- COVID19n$cell.info
# geneObject <- estPreParas.multi(counts, cell.info,
#                     id="SampleId", x1="condition",
#                     cells.interesting=c("T cells", "DC", "Prolif.T")[c(2,3)])
# Genes.tested <- geneCandidate(geneObject)
# m.n.power <- sizeCal.multi(low.up.m=c(10,14), low.up.n=c(200,500),
#     ePower=0.8, FDR=0.05, grid.m=1, grid.n=50, r=1, rc=1, total=NULL, Genes.tested)
# head(optimalCost(m.n.power))
}
