% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPower.sep.R
\name{plotPower.sep}
\alias{plotPower.sep}
\title{Figure Showing Numbers of Cells Required by Cell Types of Interest}
\usage{
plotPower.sep(
  m.n.power,
  ct.names = NULL,
  ePower = 0.8,
  nrow = NULL,
  ncol = NULL
)
}
\arguments{
\item{m.n.power}{Object of sizeCal.multi or sizeCal.multi.BA.}

\item{ct.names}{Names of cell types.}

\item{ePower}{Expected overall power (expected proportion of DEGs detected).}

\item{nrow}{Number of rows to show figures.}

\item{ncol}{Number of columns to show figures.}
}
\description{
A function to plot numbers of cells required by cell types of interest.
}
\examples{
# https://github.com/cyhsuTN/scPS
# load(file = "DataForDemo/COVID19n.rda")
# counts <- COVID19n$counts
# cell.info <- COVID19n$cell.info
# geneObject <- estPreParas.multi(counts, cell.info,
#                     id="SampleId", x1="condition",
#                     cells.interesting=c("T cells", "DC", "Prolif.T")[c(2,3)])
# Genes.tested <- geneCandidate(geneObject)
# m.n.power <- sizeCal.multi(low.up.m=c(10,14), low.up.n=c(200,500),
#     ePower=0.8, FDR=0.05, grid.m=1, grid.n=50, r=1, rc=1, total=NULL, Genes.tested)
# plotPower.sep(m.n.power)
}
