% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerCal.BA.R
\name{powerCal.BA}
\alias{powerCal.BA}
\title{Power Calculation for DEA of a Single Cell Type between Paired Groups}
\usage{
powerCal.BA(ns, m, vvmean1, FC, vvrho, hf, FDR)
}
\arguments{
\item{ns}{Numbers of cells of interest in two stages per subject.}

\item{m}{Number of subjects.}

\item{vvmean1}{Gene means of candidate genes in the pre-stage.}

\item{FC}{Fold changes of candidate genes between two stages.}

\item{vvrho}{Cell-cell correlations of candidate genes within subjects.}

\item{hf}{A curve showing the relationship between gene means and gene standard deviations.}

\item{FDR}{False discovery rate.}
}
\value{
\item{power}{Statistical power}

\item{alpha}{Type I error for each gene comparison}
}
\description{
A function to calculate powers for DEA of a single cell type between paired groups.
}
\examples{
set.seed(12345)
vvmean1 <- rgamma(1000, shape=2, scale=0.5)
FC <- c(rep(2, 50), rep(1, 950))
ab <- gammaTrans(mean=0.01, q95=0.1) # Output shape and scale parameters
vvrho <- rgamma(1000, shape=ab[1], scale=ab[2])
hf <- function(x) sqrt(x*(1+3*x))
powerCal.BA(ns=c(1,1)*70, m=8, vvmean1, FC, vvrho, hf, FDR=0.05)
}
