% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerCal.multi.BA.R
\name{powerCal.multi.BA}
\alias{powerCal.multi.BA}
\title{Power Calculation for DEA of Multiple Cell Types between Paired Groups}
\usage{
powerCal.multi.BA(ns = c(600, 600), m = 10, Genes.tested, FDR)
}
\arguments{
\item{ns}{Numbers of total cells in two stages per subject.}

\item{m}{Number of subjects.}

\item{Genes.tested}{Object of estPreParas.multi.}

\item{FDR}{False discovery rate.}
}
\description{
A function to calculate powers for DEA of multiple cell types between paired groups.
}
\examples{
# https://github.com/cyhsuTN/scPS
# load(file = "DataForDemo/GSE120575n.rda")
# counts <- GSE120575n$counts
# cell.info <- GSE120575n$cell.info
# cell.info$TX <- factor(cell.info$TX, levels = c("Pre", "Post"))
# geneObject <- estPreParas.multi(counts, cell.info, id="ptID", x1="TX",
#                     cells.interesting=c("NK", "B"))
# Genes.tested <- geneCandidate(geneObject)
# powerCal.multi.BA(ns=c(1,1)*550, m=11, Genes.tested, FDR=0.05)
}
