% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerCal.multi.R
\name{powerCal.multi}
\alias{powerCal.multi}
\title{Power Calculation for DEA of Multiple Cell Types between Two Independent Groups}
\usage{
powerCal.multi(ns, ms, Genes.tested, FDR)
}
\arguments{
\item{ns}{Number of total cells required per subject in control and experimental groups.}

\item{ms}{Numbers of subjects in control and experimental groups, respectively.}

\item{Genes.tested}{Object of estPreParas.multi.}

\item{FDR}{False discovery rate.}
}
\description{
A function to calculate powers for DEA of multiple cell types
between two independent groups.
}
\examples{
# https://github.com/cyhsuTN/scPS
# load(file = "DataForDemo/COVID19n.rda")
# counts <- COVID19n$counts
# cell.info <- COVID19n$cell.info
# geneObject <- estPreParas.multi(counts, cell.info,
#                     id="SampleId", x1="condition",
#                     cells.interesting=c("T cells", "DC", "Prolif.T")[c(2,3)])
# Genes.tested <- geneCandidate(geneObject)
# powerCal.multi(ns=c(400,400), ms=c(1,1)*11, Genes.tested, FDR=0.05)
}
