% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerCal.R
\name{powerCal}
\alias{powerCal}
\title{Power Calculation for DEA of a Single Cell Type between Two Independent Groups}
\usage{
powerCal(ns, ms, vvmean1, FC, vvrho, hf, FDR)
}
\arguments{
\item{ns}{Number of cells of interest per subject in control and experimental groups.}

\item{ms}{Numbers of subjects in the control and experimental groups, respectively.}

\item{vvmean1}{Gene means of candidate genes in the control group.}

\item{FC}{Fold changes of candidate genes between the control and the experimental groups.}

\item{vvrho}{Cell-cell correlations of candidate genes within subjects.}

\item{hf}{A curve showing the relationship between gene means and gene standard deviations.}

\item{FDR}{False discovery rate.}
}
\value{
\item{power}{Statistical power}

\item{alpha}{Type I error for each gene comparison}
}
\description{
A function to calculate powers for DEA of a single cell type between two independent groups.
}
\examples{
set.seed(12345)
vvmean1 <- rep(1, 1000)
FC <- c(rep(2, 50), rep(1, 950))
ab <- gammaTrans(mean=0.01, q95=0.1) # Output shape and scale parameters
vvrho <- rgamma(1000, shape=ab[1], scale=ab[2])
hf <- function(x) sqrt(x*(1+3*x))
powerCal(ns=c(30,50), ms=c(1,1)*10, vvmean1, FC, vvrho, hf, FDR=0.05)
}
