% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeCal.BA.R
\name{sizeCal.BA}
\alias{sizeCal.BA}
\title{Sample Size Calculation for DEA of a Single cell Type between Paired Groups}
\usage{
sizeCal.BA(
  low.up.m = c(2, 10),
  low.up.n = c(10, 100),
  ePower = 0.8,
  FDR = 0.05,
  grid.m = 2,
  grid.n = 10,
  rc = 1,
  vvmean1,
  FC,
  vvrho,
  hf = function(x) sqrt(x)
)
}
\arguments{
\item{low.up.m}{Lower and upper bounds of subjects required.}

\item{low.up.n}{Lower and upper bounds of cells of interest required in the pre-stage per subject.}

\item{ePower}{An expected overall power (an expected proportion of DEGs detected).}

\item{FDR}{False discovery rate.}

\item{grid.m}{An integer by which the \emph{low.up.m} is divided.}

\item{grid.n}{An integer by which the \emph{low.up.n} is divided.}

\item{rc}{A ratio of no. of cells in post-stage to in pre-stage.}

\item{vvmean1}{Gene means of candidate genes in the pre-stage.}

\item{FC}{Fold changes of candidate genes between two stages.}

\item{vvrho}{Cell-cell correlations of candidate genes within subjects.}

\item{hf}{A curve showing the relationship between gene means and gene standard deviations.}
}
\value{
\item{m.n.power}{A dataframe showing powers}

\item{fig}{A heat map}
}
\description{
A function to calculate sample size and numbers of cells of interest
between paired groups.
}
\examples{
set.seed(12345)
vvmean1 <- rgamma(1000, shape=2, scale=0.5)
FC <- c(rep(2, 50), rep(1, 950))
ab <- gammaTrans(mean=0.01, q95=0.1) # Output shape and scale parameters
vvrho <- rgamma(1000, shape=ab[1], scale=ab[2])
hf <- function(x) sqrt(x*(1+3*x))
sizeCal.BA(low.up.m=c(9,13), low.up.n=c(30,60), ePower=0.8, FDR=0.05,
           grid.m=1, grid.n=5, rc=1, vvmean1, FC, vvrho, hf)
}
