% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeCal.multi.BA.R
\name{sizeCal.multi.BA}
\alias{sizeCal.multi.BA}
\title{Sample Size Calculation for DEA of Multiple Cell Types between Paired Groups}
\usage{
sizeCal.multi.BA(
  low.up.m = c(2, 10),
  low.up.n = c(10, 100),
  ePower = 0.8,
  FDR = 0.05,
  grid.m = 2,
  grid.n = 10,
  rc = 1,
  Genes.tested
)
}
\arguments{
\item{low.up.m}{Lower and upper bounds of subjects required.}

\item{low.up.n}{Lower and upper bounds of total cells required in the pre-stage per subject.}

\item{ePower}{An expected overall power (an expected proportion of DEGs detected).}

\item{FDR}{False discovery rate.}

\item{grid.m}{An integer by which the \emph{low.up.m} is divided.}

\item{grid.n}{An integer by which the \emph{low.up.n} is divided.}

\item{rc}{A ratio of no. of cells in post-stage to no. of cells in pre-stage.}

\item{Genes.tested}{Object of estPreParas.multi.}
}
\description{
A function to calculate sample sizes and numbers of total cells
between paired groups.
}
\examples{
# https://github.com/cyhsuTN/scPS
# load(file = "DataForDemo/GSE120575n.rda")
# counts <- GSE120575n$counts
# cell.info <- GSE120575n$cell.info
# cell.info$TX <- factor(cell.info$TX, levels = c("Pre", "Post"))
# geneObject <- estPreParas.multi(counts, cell.info, id="ptID", x1="TX",
#                     cells.interesting=c("NK", "B"))
# Genes.tested <- geneCandidate(geneObject)
# m.n.power <- sizeCal.multi.BA(low.up.m=c(10,14), low.up.n=c(400,700),
#     ePower=0.8, FDR=0.05, grid.m=1, grid.n=50, rc=1, Genes.tested)
# m.n.power$fig

}
