% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeCal.multi.R
\name{sizeCal.multi}
\alias{sizeCal.multi}
\title{Sample Size Calculation for DEA of Multiple Cell Types between Two Independent Groups}
\usage{
sizeCal.multi(
  low.up.m = c(2, 10),
  low.up.n = c(10, 100),
  ePower = 0.8,
  FDR = 0.05,
  grid.m = 2,
  grid.n = 10,
  r = 1,
  rc = 1,
  total = NULL,
  Genes.tested
)
}
\arguments{
\item{low.up.m}{Lower and upper bounds of subjects in control.}

\item{low.up.n}{Lower and upper bounds of total cells per subject in control.}

\item{ePower}{An expected overall power (an expected proportion of DEGs detected).}

\item{FDR}{False discovery rate.}

\item{grid.m}{An integer by which the \emph{low.up.m} is divided.}

\item{grid.n}{An integer by which the \emph{low.up.n} is divided.}

\item{r}{An allocation ratio (a ratio of no. of subjects in experimental
to no. of subjects in control).}

\item{rc}{A ratio of no. of cells in experimental to in control.}

\item{total}{A total sample size.}

\item{Genes.tested}{Object of estPreParas.multi.}
}
\description{
A function to calculate sample sizes and numbers of total cells
between two independent groups.
}
\examples{
# https://github.com/cyhsuTN/scPS
# load(file = "DataForDemo/COVID19n.rda")
# counts <- COVID19n$counts
# cell.info <- COVID19n$cell.info
# geneObject <- estPreParas.multi(counts, cell.info,
#                     id="SampleId", x1="condition",
#                     cells.interesting=c("T cells", "DC", "Prolif.T")[c(2,3)])
# Genes.tested <- geneCandidate(geneObject)
# m.n.power <- sizeCal.multi(low.up.m=c(10,14), low.up.n=c(200,500),
#     ePower=0.8, FDR=0.05, grid.m=1, grid.n=50, r=1, rc=1, total=NULL, Genes.tested)
# m.n.power$fig
}
