% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeCal.R
\name{sizeCal}
\alias{sizeCal}
\title{Sample Size Calculation for DEA of a Single cell Type between Two Independent Groups}
\usage{
sizeCal(
  low.up.m = c(6, 10),
  low.up.n = c(10, 100),
  ePower = 0.8,
  FDR = 0.05,
  grid.m = 1,
  grid.n = 10,
  r = 1,
  rc = 1,
  total = NULL,
  vvmean1,
  FC,
  vvrho,
  hf = function(x) sqrt(x)
)
}
\arguments{
\item{low.up.m}{Lower and upper bounds of subjects in control.}

\item{low.up.n}{Lower and upper bounds of cells of interest in each subject of control.}

\item{ePower}{An expected overall power (an expected proportion of DEGs detected).}

\item{FDR}{False discovery rate.}

\item{grid.m}{An integer by which the \emph{low.up.m} is divided.}

\item{grid.n}{An integer by which the \emph{low.up.n} is divided.}

\item{r}{An allocation ratio (a ratio of no. of experimental to no. of control).}

\item{rc}{A ratio of no. of cells in experimental to in control.}

\item{total}{A total of sample size.}

\item{vvmean1}{Gene means of candidate genes in the control group.}

\item{FC}{Fold changes of candidate genes between the control and the experimental groups.}

\item{vvrho}{Cell-cell correlations of candidate genes within subjects.}

\item{hf}{A curve showing the relationship between gene means and gene deviations.}
}
\value{
\item{m.n.power}{A dataframe showing powers}

\item{fig}{A heat map}
}
\description{
A function to calculate sample sizes and numbers of cells of interest
between two independent groups.
}
\examples{
set.seed(12345)
vvmean1 <- rep(1, 1000)
FC <- c(rep(2, 50), rep(1, 950))
ab <- gammaTrans(mean=0.01, q95=0.1) # Output shape and scale parameters
vvrho <- rgamma(1000, shape=ab[1], scale=ab[2])
hf <- function(x) sqrt(x*(1+3*x))
sizeCal(low.up.m=c(7,11), low.up.n=c(30,60), ePower=0.8, FDR=0.05,
       grid.m=1, grid.n=5, r=1, rc=1, total=NULL, vvmean1, FC, vvrho, hf)
}
